python/requirements.txt
Appium-Python-Client==1.1.0
requests==2.25.1